requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
langchain-classic
langchain-openai
chroma-hnswlib
chromadb
langchain-text-splitters
tiktoken
langchain-community
